+++ cypress/tests/core/blocks-listing.js
@@ -188,7 +188,7 @@ describe('Listing Block Tests', () => {
       .contains('Page')
       .click();
 
-    //before save, vrify if in list there's a page with id my-page-test
+    //before save, verify if in list there's a page with id my-page-test
     cy.get(`.block.listing .listing-body:first-of-type`).contains('My Page');
     //before save, verify if in list there isn't the News with title My News
     cy.get(`.block.listing .listing-body`)
@@ -475,6 +475,7 @@ describe('Listing Block Tests', () => {
       .contains('Document outside Folder')
       .should('not.exist');
   });
+
   it('Listing block - Test Criteria: Location relative with some outside content', () => {
     // Given we have two document about us, contact at portal route and two document in My Page
     // i.e News Item One and News Item Two
@@ -555,6 +556,94 @@ describe('Listing Block Tests', () => {
     cy.get(`.block.listing .listing-body`)
       .contains('about us')
       .should('not.exist');
+  });
+
+  it('Listing block: respect batching and limits', () => {
+    // Given One Document My Page Test and One News Item MY News and One Folder My Folder
+    cy.createContent({
+      contentType: 'Document',
+      contentId: 'my-page-test',
+      contentTitle: 'My Page Test',
+      path: 'my-page',
+    });
+    cy.createContent({
+      contentType: 'News Item',
+      contentId: 'my-news',
+      contentTitle: 'My News',
+      path: 'my-page',
+    });
+    cy.createContent({
+      contentType: 'Folder',
+      contentId: 'my-folder',
+      contentTitle: 'My Folder',
+      path: 'my-page',
+    });
+
+    cy.visit('/my-page');
+    cy.waitForResourceToLoad('@navigation');
+    cy.waitForResourceToLoad('@breadcrumbs');
+    cy.waitForResourceToLoad('@actions');
+    cy.waitForResourceToLoad('@types');
+    cy.waitForResourceToLoad('my-page');
+    cy.navigate('/my-page/edit');
+
+    cy.get(`.block.title [data-contents]`)
+      .clear()
+      .type('Listing block - respect batching and limits');
+
+    //add listing block
+    cy.get('.block.text [contenteditable]').click();
+    cy.get('button.block-add-button').click();
+    cy.get('.blocks-chooser .title').contains('Common').click();
+    cy.get('.blocks-chooser .common').contains('Listing').click();
+
+    //verify before save
+    cy.get(`.block.listing .listing-body:first-of-type`).contains(
+      'My Page Test',
+    );
+
+    cy.get('.sidebar-container .tabs-wrapper .menu .item')
+      .contains('Block')
+      .click();
+    cy.get('.querystring-widget .fields').contains('Add criteria').click();
+    cy.get(
+      '.querystring-widget .fields:first-of-type .field:first-of-type .react-select__menu .react-select__option',
+    )
+      .contains('Location')
+      .click();
+    //location relative..
+    cy.get(
+      '.querystring-widget .fields:first-of-type .main-fields-wrapper .field:last-of-type',
+    ).click();
+
+    cy.get(
+      '.querystring-widget .fields:first-of-type .main-fields-wrapper .field:last-of-type .react-select__menu .react-select__option',
+    )
+      .contains('Absolute path')
+      .click();
+    cy.get('.twelve > :nth-child(1) > :nth-child(2) > .ui > input').type('/');
+
+    cy.get('#field-limit-4-querystring').click().type('2');
+
+    //save
+    cy.get('#toolbar-save').click();
+
+    //test after save
+    cy.get('.listing-item').should(($els) => {
+      expect($els).to.have.length(2);
+    });
+
+    cy.navigate('/my-page/edit');
+    cy.get('.block-editor-listing').click();
+    cy.get('.sidebar-container .tabs-wrapper .menu .item')
+      .contains('Block')
+      .click();
+
+    cy.get('#field-limit-4-querystring').click().clear().type('0');
+    cy.get('#field-batch_size-5-querystring').click().type('2');
+    cy.get('.ui.pagination.menu a[value="2"]').first().click();
+
+    cy.get('.listing-item h4').first().contains('My Folder');
   });
 
   it('Listing block - Test Criteria: Location Navigation', () => {